Level up system and stats increase on level up
Created 2 events, one for LevelUp and the other one for ExpChange, both of them connected to the update slider (which needs to be assigned when ready).

Player currently increases max health, max shield and base armor on level up.

If you press L key ingame you can debug the level system and the stats increase on level up.