Auto Aim
Script EnemyRadar added, it will find the closest enemy inside a trigger collision
and return the game object to player controller, which if true will replace the manual rotation with the coordinate of the colese enemy.

in player manager added the code "  if (other.tag == "ColectableItem")" in line 129 so that the colision trigger will not start the colectible pick up.