Auto Aim
Auto aim working with raycast and bullet weapon
it seaches the closest and if any obstacle in front will not shoot.

the main script is "FieldOFView", "FieldOFViewEditor" just displays a gui in editor.

enemy radar script is deleted.

Muzzle flash spawn was removed from projectile script and  placed in ProjectileManager, it was spawned only with bullet now is shared betewn any of  the two tipes of weapon shooting.

as well, projectile manager  only activates weapon shooting if there is any enemy in sight

player controller auto aims if there is any enemy close by